% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.L0TFinvfix}
\alias{coef.L0TFinvfix}
\alias{coef.L0TFinvopt}
\title{Extract estimated trends}
\usage{
\method{coef}{L0TFinvfix}(object, k = NULL)

\method{coef}{L0TFinvopt}(object, k = NULL)
}
\arguments{
\item{object}{The output of L0TFinvfix or L0TFinvopt}

\item{k}{The given number of change points}
}
\description{
Extract the coefficients of the estimated trends under the constraint of a given number of change points.
}
\seealso{
\code{\link{print.L0TFinvfix}}
}
