% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sMatrix.R
\name{DiffMat}
\alias{DiffMat}
\title{Generate a difference matrix}
\usage{
DiffMat(n, q)
}
\arguments{
\item{n}{The number of data points}

\item{q}{The order of the difference}
}
\value{
A matrix with dimensions \eqn{n-q-1} by \eqn{n}, whose elements correspond to the combinatorial values of \eqn{q}.
}
\description{
This function generates a matrix for computing differences of a certain order, useful in numerical methods and for creating specific matrix patterns.
}
\examples{
Mat1 <- DiffMat(n = 10, q = 0)
print(Mat1)

Mat2 <- DiffMat(n = 15, q = 1)
print(Mat2)

Mat3 <- DiffMat(n = 15, q = 2)
print(Mat3)

}
\seealso{
\code{\link{XMat}}
}
